requirements.txt
anyio==3.7.0
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
databases==0.7.0
exceptiongroup==1.1.1
fastapi==0.97.0
Genderize==0.3.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
numpy==1.25.0
pandas==2.0.3
psycopg2-binary==2.9.6
pydantic==1.10.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.5
uvicorn==0.22.0
XlsxWriter==3.1.2
